split out magic command completion tests
